+++ app/Http/Requests/Api/Application/Servers/ServerWriteRequest.php
@@ -2,7 +2,6 @@
 
 namespace Pterodactyl\Http\Requests\Api\Application\Servers;
 
-use Pterodactyl\Models\Server;
 use Pterodactyl\Services\Acl\Api\AdminAcl;
 use Pterodactyl\Http\Requests\Api\Application\ApplicationApiRequest;
 
@@ -17,16 +16,4 @@ class ServerWriteRequest extends ApplicationApiRequest
      * @var int
      */
     protected $permission = AdminAcl::WRITE;
-
-    /**
-     * Determine if the requested server exists on the Panel.
-     *
-     * @return bool
-     */
-    public function resourceExists(): bool
-    {
-        $server = $this->route()->parameter('server');
-
-        return $server instanceof Server && $server->exists;
-    }
 }